+++ CourtDatabase2/Controllers/DebitorsController.cs
@@ -144,8 +144,12 @@ namespace CourtDatabase2.Controllers
         // POST: Debitors/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
+        public async Task<IActionResult> DeleteConfirmed(int? id)
         {
+            if (id == null)
+            {
+                return NotFound();
+            }
             var debitor = await dbContext.Debitors.FindAsync(id);
             dbContext.Debitors.Remove(debitor);
             await dbContext.SaveChangesAsync();
