+++ CourtDatabase2/Controllers/ExecutorsController.cs
@@ -33,38 +33,73 @@ namespace CourtDatabase2.Controllers
         [HttpPost]
         public IActionResult Create(ExecutorsCreateViewModel model)
         {
+            if (!ModelState.IsValid)
+            {
+                return this.View(model);
+            }
             this.executorService.Create(model);
             return this.RedirectToAction("All");
         }
 
         public IActionResult Details(int? id)
         {
+            if (id == null)
+            {
+                return NotFound();
+            }
             var viewModel = this.executorService.Details(id);
+            if (viewModel == null)
+            {
+                return NotFound();
+            }
             return this.View(viewModel);
         }
 
         public IActionResult Edit(int? id)
         {
+            if (id == null)
+            {
+                return NotFound();
+            }
             var viewModel = this.executorService.Details(id);
+            if (viewModel == null)
+            {
+                return NotFound();
+            }
             return this.View(viewModel);
         }
 
         [HttpPost]
         public IActionResult Edit(ExecutorsEditViewModel model)
         {
+            if (!ModelState.IsValid)
+            {
+                return this.View(model);
+            }
             this.executorService.Edit(model);
-            
             return this.RedirectToAction("All");
         }
 
         public IActionResult Delete(int? id)
         {
+            if (id == null)
+            {
+                return NotFound();
+            }
             var viewModel = this.executorService.Details(id);
+            if (viewModel == null)
+            {
+                return NotFound();
+            }
             return this.View(viewModel);
         }
 
         public IActionResult DeleteConfirm(int? id)
         {
+            if (id == null)
+            {
+                return NotFound();
+            }
             this.executorService.Delete(id);
             return this.RedirectToAction("All");
         }
